add hooks file for workspaceitemstable
